==> 2_Multivariate analysis.ipynb <==
# Multivariate analysis

Multivariate analysis serves as a powerful toolkit enabling us to derive deeper insights where multiple variables interact. By examining the relationships and patterns among several variables simultaneously, this approach enhances our understanding of the data and helps us find hidden connections, identifying underlying factors that drive observed trends, and making more informed decisions.

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import pyMechkar
import sweetviz as sv
#import pandas_profiling
from ydata_profiling import ProfileReport
from scipy.stats import chi2_contingency
from matplotlib.backends.backend_pdf import PdfPages
from scipy.stats import spearmanr
import matplotlib.backends.backend_pdf


Read the data

df = pd.read_csv('../sql/home_credit_train_ff.csv', index_col=0)

As before, divide the feautes into numerical and categorical types

cat_cols = []
num_cols = []

for col in df.columns:
    if df[col].dtype == 'object' or pd.api.types.is_categorical_dtype(df[col]) or df[col].nunique() <= 8:
        cat_cols.append(col)
    else:
        num_cols.append(col)

print(f'string or categorical columns:\n {cat_cols}')
print('--------------------------------------------------------------------------------------------------------------------')
print(f'numeric columns:\n {num_cols}')


We will display the numerical columns in a scatter plots

# Create a PDF file for each variable
for variable in num_cols:
    pdf_filename = f'{variable}_plots.pdf'
    with PdfPages(pdf_filename) as pdf:
        plt.figure(figsize=(6, 4))
        for other_variable in num_cols:
            if other_variable != variable:
                plt.scatter(df[variable], df[other_variable], color='blue', alpha=0.7, rasterized=True)
                plt.xlabel(variable)
                plt.ylabel(other_variable)
                plt.title(f'{variable} vs {other_variable}')
                plt.grid(True)
                pdf.savefig()  # Save the current figure to the PDF
                plt.close()

print("PDF files with plots created.")


And the categorical columns using countplots

# Create a PDF file for each variable
for variable in cat_cols:
    pdf_filename = f'{variable}_plots.pdf'
    with PdfPages(pdf_filename) as pdf:
        plt.figure(figsize=(6, 4))
        for other_variable in cat_cols:
            if other_variable != variable:
                #plt.scatter(df[variable], df[other_variable], color='blue', alpha=0.7, rasterized=True)
                sns.countplot(data=df, x=variable, hue=other_variable, rasterized=True)
                plt.xlabel(variable)
                plt.ylabel('Count')
                plt.title(f'{variable} vs {other_variable}')
                plt.grid(True)
                pdf.savefig()  # Save the current figure to the PDF
                plt.legend(title=other_variable)
                plt.xticks(rotation=45)
                plt.close()

print("PDF files with plots created.")

pdf = PdfPages(r'countplot cat_cols vs target.pdf')

for col in cat_cols:
    if col != 'TARGET':
        plt.figure(figsize=(4,3))
        sns.countplot(data=df, x=col, hue='TARGET')
        plt.xlabel(col)
        plt.ylabel('Count')
        plt.title('countplot - ' + col + ' ' + 'vs TARGET')
        plt.legend(title='TARGET')
        plt.xticks(rotation=45)
        pdf.savefig(bbox_inches='tight', dpi=300)
        plt.show()
    
pdf.close()

conclusions from the plots that were created above will appear in the next notebook
