==> 8_Split and data balance.ipynb <==
# Split and Data Balance

import pandas as pd
import numpy as np
pd.set_option('display.max_rows', None) 


df = pd.read_csv('df_after_feature_engineering.csv', index_col=0)

#%store -r df1

#df = df1

df.head()

df.shape

print(df.dtypes)

for column, data_type in df.dtypes.iteritems():
    if data_type == 'category':
        df[column] = df[column].astype(int)

print(df.dtypes)

df.shape

The primary evaluation metric for the model will be recall.
Recall is vital in this context as it focuses on minimizing false negatives, ensuring that instances of payment difficulties are not overlooked. A missed instance of payment difficulty (false negative) could have more severe consequences than a false positive, making recall the ideal metric to prioritize in this loan prediction scenario.

# Split Data

split the dataset into three parts:

Test Set (20%): This set remains untouched until the very end and helps handle the model's real-world performance.

Validation Set (16%): A subset of the data used to fine-tune the model's parameters and prevent overfitting during development.

Training Set (64%): The largest portion where the model learns patterns and features to make accurate predictions.

## remove unnecessary vars
X = df.loc[:,df.columns!='TARGET']

X.head()

y = df.loc[:,df.columns=='TARGET']
y.head()

from sklearn.model_selection import train_test_split

#split to train and test -we won't use the test until the end
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)


#split to train and val 
X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.2, random_state=42)



X_test.shape

X_train.shape

X_val.shape

y_test.shape

y_train.shape

y_val.shape

y_test.value_counts()

y_train.value_counts()

y_val.value_counts()

# Imbalance Data 

we're addressing class imbalance in a classification dataset using the imblearn library. 

The goal is to balance the distribution of data to improve model performance. 

Each function takes input data (X) and labels (y) and returns balanced samples.
These techniques aim to equip the model with more equitable data for accurate classification, mitigating class imbalance issues

# !pip install imbalanced-learn

from imblearn.under_sampling import RandomUnderSampler
from imblearn.over_sampling import RandomOverSampler
from imblearn.combine import SMOTEENN # Combination of SMOTE and Edited Nearest Neighbors
from imblearn.over_sampling import SMOTE

def perform_under_sampling(X, y, random_state=None):
    under_sampler = RandomUnderSampler( random_state=random_state)
    X_resampled, y_resampled = under_sampler.fit_resample(X, y)
    return X_resampled, y_resampled

def perform_over_sampling(X, y, random_state=None):
    over_sampler = RandomOverSampler( random_state=random_state)
    X_resampled, y_resampled = over_sampler.fit_resample(X, y)
    return X_resampled, y_resampled

def perform_overunder_sampling(X, y, random_state=None):
    combined_sampler = SMOTEENN( random_state=random_state)
    X_resampled, y_resampled = combined_sampler.fit_resample(X, y)
    return X_resampled, y_resampled

def perform_SMOTE_sampling(X, y, random_state=None):
    combined_sampler = SMOTE(random_state=random_state)
    X_resampled, y_resampled = combined_sampler.fit_resample(X, y)
    return X_resampled, y_resampled


X = X_train
y = y_train

# Under-sampling
X_under, y_under = perform_under_sampling(X, y, random_state=42)

# Over-sampling
X_over, y_over = perform_over_sampling(X, y,  random_state=42)

# Combined SMOTE-sampling
X_overunder, y_overunder = perform_overunder_sampling(X, y, random_state=42)

# Combined SMOTE-sampling
X_SMOTE, y_SMOTE = perform_SMOTE_sampling(X, y, random_state=42)


X_under.shape

X_over.shape

X_overunder.shape

X_SMOTE.shape

we choose random forest and RECALL and see which dataset is the best

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score


# Create a Random Forest classifier
rf_classifier = RandomForestClassifier(n_estimators=100, random_state=42)

# Train the classifier on the training data
rf_classifier.fit(X_train, y_train)

# Make predictions on the validation data
y_val_pred = rf_classifier.predict(X_val)
y_train_pred = rf_classifier.predict(X_train)

# Calculate overall recall
overall_val_recall = recall_score(y_val, y_val_pred, average='weighted')
overall_train_recall = recall_score(y_train, y_train_pred, average='weighted')
print("overall_val_recall:", overall_val_recall)
print("overall_train_recall:", overall_train_recall)

ratio_rf = overall_val_recall/ overall_train_recall
print("ratio val/ train:", ratio_rf)

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score


# Create a Random Forest classifier
rf_classifier = RandomForestClassifier(n_estimators=100, random_state=42)

# Train the classifier on the training data
rf_classifier.fit(X_under, y_under)

# Make predictions on the validation data
y_val_pred = rf_classifier.predict(X_val)
y_train_pred = rf_classifier.predict(X_under)

# Calculate overall recall
overall_val_recall = recall_score(y_val, y_val_pred, average='weighted')
overall_train_recall = recall_score(y_under, y_train_pred, average='weighted')
print("overall_val_recall:", overall_val_recall)
print("overall_train_recall:", overall_train_recall)

ratio_rf = overall_val_recall/ overall_train_recall
print("ratio val/ train:", ratio_rf)

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score


# Create a Random Forest classifier
rf_classifier = RandomForestClassifier(n_estimators=100, random_state=42)

# Train the classifier on the training data
rf_classifier.fit(X_over, y_over)

# Make predictions on the validation data
y_val_pred = rf_classifier.predict(X_val)
y_train_pred = rf_classifier.predict(X_over)

# Calculate overall recall
overall_val_recall = recall_score(y_val, y_val_pred, average='weighted')
overall_train_recall = recall_score(y_over, y_train_pred, average='weighted')
print("overall_val_recall:", overall_val_recall)
print("overall_train_recall:", overall_train_recall)

ratio_rf = overall_val_recall/ overall_train_recall
print("ratio val/ train:", ratio_rf)

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score



# Create a Random Forest classifier
rf_classifier = RandomForestClassifier(n_estimators=100, random_state=42)

# Train the classifier on the training data
rf_classifier.fit(X_overunder, y_overunder)

# Make predictions on the validation data
y_val_pred = rf_classifier.predict(X_val)
y_train_pred = rf_classifier.predict(X_overunder)

# Calculate overall recall
overall_val_recall = recall_score(y_val, y_val_pred, average='weighted')
overall_train_recall = recall_score(y_overunder, y_train_pred, average='weighted')
print("overall_val_recall:", overall_val_recall)
print("overall_train_recall:", overall_train_recall)

ratio_rf = overall_val_recall/ overall_train_recall
print("ratio val/ train:", ratio_rf)


from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score



# Create a Random Forest classifier
rf_classifier = RandomForestClassifier(n_estimators=100, random_state=42)

# Train the classifier on the training data
rf_classifier.fit(X_SMOTE, y_SMOTE)


# Make predictions on the validation data
y_val_pred = rf_classifier.predict(X_val)
y_train_pred = rf_classifier.predict(X_SMOTE)

# Calculate overall recall
overall_val_recall = recall_score(y_val, y_val_pred, average='weighted')
overall_train_recall = recall_score(y_SMOTE, y_train_pred, average='weighted')
print("overall_val_recall:", overall_val_recall)
print("overall_train_recall:", overall_train_recall)

ratio_rf = overall_val_recall/ overall_train_recall
print("ratio val/ train:", ratio_rf)

Selecting the model with the highest ratio of validation recall to training recall that isn't overfitting. This approach involves comparing how well the model performs on the validation data relative to the training data and choosing the one that strikes the best balance between generalization and overfitting.

X_train = X_over
y_train = y_over

%store X_train
%store y_train

%store X_test
%store y_test

%store X_val
%store y_val

X_train.to_csv('X_train.csv', index=True)

y_train.to_csv('y_train.csv', index=False)

X_test.to_csv('X_test.csv', index=True)

y_test.to_csv('y_test.csv', index=False)

X_val.to_csv('X_val.csv', index=True)

y_val.to_csv('y_val.csv', index=False)
