==> 9_Models.ipynb <==
# Predictions using different models


import matplotlib.pyplot as plt
plt.style.use('classic')
%matplotlib inline

import pandas as pd
import numpy as np

from sklearn.metrics import accuracy_score

from sklearn.linear_model import LogisticRegression, RidgeClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC

import seaborn as sns
from sklearn.metrics import recall_score



%store -r X_train
%store -r y_train

%store -r X_test
%store -r y_test

%store -r X_val
%store -r y_val

def custom_recall(y_true, y_pred):
    return recall_score(y_true, y_pred, average='weighted')

print(X_train.shape)
X_train.head()

X_train.dtypes

# initiate the model's selection dataframe
res = pd.DataFrame()

# Linear Regression (Ridge)

mod1 = LogisticRegression(penalty='l2')
mod1.fit(X_train,y_train)

mod1

pred1 = mod1.predict(X_val)

res = res.append({'model':'Logistic Regression (Ridge)','custom_recall':custom_recall(y_val,pred1)},ignore_index=True)

res

# LASSO

mod3 = LogisticRegression(penalty='l1',solver= 'liblinear', random_state=10)
mod3.fit(X_train,y_train)
mod3

pred3 = mod3.predict(X_val)

res = res.append({'model':'LASSO','custom_recall':custom_recall(y_val,pred3)},ignore_index=True)

## Decision Trees

mod5 = DecisionTreeClassifier(random_state=2)
mod5.fit(X_train,y_train)

pred5 = mod5.predict(X_val)

res = res.append({'model':'Decision Tree','custom_recall':custom_recall(y_val,pred5)},ignore_index=True)

res

## Random Forest

mod6 = RandomForestClassifier(random_state=3)

mod6.fit(X_train,y_train)

pred6 = mod6.predict(X_val)

res = res.append({'model':'Random Forest','custom_recall':custom_recall(y_val,pred6)},ignore_index=True)

res

## Adaptive Boosting (ADABoost)

mod7 = AdaBoostClassifier(random_state=1)

mod7.fit(X_train,y_train)

pred7 = mod7.predict(X_val)

res = res.append({'model':'AdaBoost','custom_recall':custom_recall(y_val,pred7)},ignore_index=True)

res

## Gradient Boosting Machine

mod8 = GradientBoostingClassifier(random_state=2)

mod8.fit(X_train,y_train)

pred8 = mod8.predict(X_val)

res = res.append({'model':'Gradient Boosting','custom_recall':custom_recall(y_val,pred8)},ignore_index=True)

res

# kNN


mod9 = KNeighborsClassifier()


mod9.fit(X_train,y_train)

pred9 = mod9.predict(X_val.values)

res = res.append({'model':'kNN','custom_recall':custom_recall(y_val,pred9)},ignore_index=True)

res

# SVC



mod10 = SVC()
mod10.fit(X_train,y_train)

pred10 = mod10.predict(X_val)

res = res.append({'model':'SVC','custom_recall':custom_recall(y_val,pred10)},ignore_index=True)

# Model Selection

res = res.sort_values('custom_recall')
print(res)

Based on the custom_recall score, selecting the "Random Forest" model is a reasonable choice as it's likely to provide better recall performance for our specific use case.
