==> 10_Hyperparameters Finetuning.ipynb <==
# Hyperparameter Finetuning

import pandas as pd
import numpy as np
import seaborn as sns

from sklearn.metrics import mean_absolute_error
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn.ensemble import RandomForestClassifier

X_train = pd.read_csv('X_train.csv', index_col=0)

y_train = pd.read_csv('y_train.csv', index_col=0)

X_test = pd.read_csv('X_test.csv', index_col=0)

y_test = pd.read_csv('y_test.csv', index_col=0)

X_val = pd.read_csv('X_val.csv', index_col=0)

y_val = pd.read_csv('y_val.csv', index_col=0)

%store -r X_train
%store -r y_train

%store -r X_test
%store -r y_test

%store -r X_val
%store -r y_val

type(y_train)

y_train.value_counts()

X_train.head()

X_train.shape

X_train.isnull().sum()

X_train.dtypes

X_train.head()

## Random Search: we decide which parameters and how (randomly)

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 100, stop = 1000, num = 10)]
# Number of features to consider at every split
max_features = ['auto', 'sqrt']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 110, num = 11)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [2, 3, 5, 10]
# Minimum number of samples required at each leaf node
min_samples_leaf = [2, 3, 4, 5]
# Method of selecting samples for training each tree
bootstrap = [True, False]

# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}


print(random_grid)

rf = RandomForestClassifier()

rf_random = RandomizedSearchCV(estimator = rf, param_distributions = random_grid, n_iter = 100, cv = 5, 
                               verbose=2, random_state=42, n_jobs = -1)

# Fit the random search model
rf_random.fit(X_train, y_train)

rf_random.best_params_

from sklearn.metrics import recall_score


def evaluate(y_val, y_pred_val):
    overall_recall = recall_score(y_val, y_pred_val, average='weighted')
    print('Model Performance')
    print('Recall: {:0.4f}'.format(overall_recall))
    return overall_recall

#### Run base model


base_model = RandomForestClassifier(n_estimators = 1000, random_state = 4)
base_model.fit(X_train, y_train)
y_pred_val = base_model.predict(X_val)
#base_recall = evaluate(y_val, y_pred_val)

base_recall = evaluate(y_val, y_pred_val)

#### Compare fine-tunned model


best_random = rf_random.best_estimator_
best_random.fit(X_train, y_train)
y_pred_val = best_random.predict(X_val)
random_recall = evaluate(y_val, y_pred_val)

print('Improvement of {:0.2f}%.'.format( 100 * (base_recall - random_recall) / base_recall))

## Grid search: we decide which parameters and how (not randomly)

# Create the parameter grid 
param_grid = {
    'bootstrap': [False],
    'max_depth': [75,80,85],
    'max_features': ['sqrt', 'log2'],
    'min_samples_leaf': [2],
    'min_samples_split': [4, 5, 6],
    'n_estimators': [800, 900, 1000]
}

# Create a based model
rf = RandomForestClassifier()
# Instantiate the grid search model

grid_search = GridSearchCV(estimator = rf, param_grid = param_grid, cv = 3, n_jobs = -1, verbose = 2)

# Fit the grid search to the data
grid_search.fit(X_train, y_train)


grid_search.best_params_

best_grid = grid_search.best_estimator_
best_grid.fit(X_train, y_train)
y_pred_val = best_grid.predict(X_val)
grid_recall = evaluate(y_val, y_pred_val)

print('Improvement of {:0.2f}%.'.format( 100 * (grid_recall - grid_recall) / base_recall))

#run the test data

# TEST data

y_pred_test = base_model.predict(X_test)
base_recall = evaluate(y_test, y_pred_test)
